+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.0
+
+- Add extensions on `FormzStatus`
+- Add extensions on `FormzInputStatus`
+
 # 0.0.1
 
 Initial Release of the formz.